+++ parser/options.cc
@@ -316,6 +316,7 @@ void Options::PublishMembers()
   publish("storeTrajectoryEnergyThreshold", &Options::storeTrajectoryEnergyThreshold);
   publish("storeTrajectorySamplerID"      , &Options::storeTrajectorySamplerID);
   publish("storeTrajectoryELossSRange"    , &Options::storeTrajectoryELossSRange);
+  publish("storeSamplerAll",                &Options::storeSamplerAll);
   publish("storeSamplerCharge",             &Options::storeSamplerCharge);
   publish("storeSamplerKineticEnergy",      &Options::storeSamplerKineticEnergy);
   publish("storeSamplerMass",               &Options::storeSamplerMass);