+++ manual/source/version_history.rst
@@ -25,6 +25,7 @@ New Features
   the energy loss output. More granular than :code:`storeElossLinks`.
 * New option :code:`storeGeant4Data` to control whether the basic particle data is stored in
   the output for all particles used or not.
+* New option :code:`storeSamplerAll` to conveniently store all optional sampler data with one option.
 * Access to data version in DataLoader in analysis.
 * External geometry can be supplied as the world volume with the option :code:`worldGeometryFile`.
 * New complete physics list for crystal channelling to achieve the correct result.
@@ -139,6 +140,7 @@ Output Changes
   respectively.
 * New members in :code:`Event.Info` that are the integrated energy deposite in various parts
   for that event. These are for convenience and are the integrals of the various Eloss parts.
+* New option :code:`storeSamplerAll` to conveniently store all optional sampler data with one option.
 
 Output Classes Versions
 -----------------------